API/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-stubs==5.1.0
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.1
idna==3.10
iniconfig==2.0.0
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.0
pytest==8.3.3
requests==2.32.3
ruff==0.7.4
sqlparse==0.5.2
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
